# docs/swagger.yaml
openapi: 3.0.3
info:
  title: CoworkSpace API v2.0
  description: |
    API complete per il sistema di gestione coworking CoworkSpace v2.0.
    
    ## Caratteristiche Principali
    - **Redis Caching**: Performance ottimizzate con cache intelligente
    - **Slot Management**: Sistema automatico di gestione slot con timer
    - **Payment System**: Sistema di pagamenti semplificato
    - **Multi-role System**: Gestione utenti con ruoli diversi (utente, gestore, amministratore)
    - **RESTful API**: API REST complete per tutte le operazioni
    
    ## Autenticazione
    L'API utilizza JWT (JSON Web Tokens) per l'autenticazione. Includi il token nel header `Authorization: Bearer <token>` per le richieste protette.
    
    ## Sistema Semplificato
    Il sistema è stato semplificato rimuovendo le dipendenze real-time e i pagamenti esterni per una maggiore stabilità e facilità di deployment.
    
  version: 2.0.0
  contact:
    name: CoworkSpace Support
    email: support@coworkspace.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Server di sviluppo locale
  - url: https://coworkspace.onrender.com/api
    description: Server di produzione
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token ottenuto tramite login o registrazione
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Messaggio di errore
        code:
          type: string
          description: Codice errore specifico
      required:
        - error
...
    Spazio:
      type: object
      properties:
        id_spazio:
          type: integer
          description: ID univoco dello spazio
        nome:
          type: string
          description: Nome dello spazio
        descrizione:
          type: string
          description: Descrizione dello spazio
        id_sede:
          type: integer
          description: ID della sede
        tipo:
          type: string
          enum: [sala_riunioni, postazione_lavoro, ufficio_privato]
          description: Tipo di spazio
        prezzo_ora:
          type: number
          description: Prezzo per ora
        capacita:
          type: integer
          description: Capacità massima dello spazio
      required:
        - id_spazio
        - nome
        - id_sede
        - tipo
        - prezzo_ora
paths:
...
  /spazi:
    get:
      tags:
        - Spazi
      summary: Lista spazi disponibili
      description: Recupera la lista di tutti gli spazi disponibili
      parameters:
        - name: id_sede
          in: query
          schema:
            type: integer
          description: Filtra per sede specifica
        - name: tipo
          in: query
          schema:
            type: string
            enum: [sala_riunioni, postazione_lavoro, ufficio_privato]
          description: Filtra per tipo di spazio
      responses:
        '200':
          description: Lista spazi
          content:
            application/json:
              schema:
                type: object
                properties:
                  spazi:
                    type: array
                    items:
                      $ref: '#/components/schemas/Spazio'
        '500':
          description: Errore server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sedi:
    get:
      tags:
        - Spazi
      summary: Lista sedi
      description: Recupera la lista di tutte le sedi
      responses:
        '200':
          description: Lista sedi
          content:
            application/json:
              schema:
                type: object
                properties:
                  sedi:
                    type: array
                    items:
                      type: object
                      properties:
                        id_sede:
                          type: integer
                        nome:
                          type: string
                        indirizzo:
                          type: string
                        citta:
                          type: string
        '500':
          description: Errore server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========================================
  # PAGAMENTI
  # ========================================
...
  /admin/dashboard:
    get:
      tags:
        - Amministrazione
      summary: Dashboard amministratore
      description: Recupera dati per la dashboard amministratore
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dati dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                    description: Numero totale utenti
                  total_bookings:
                    type: integer
                    description: Numero totale prenotazioni
                  revenue_today:
                    type: number
                    description: Ricavi di oggi
                  occupancy_rate:
                    type: number
                    description: Tasso di occupazione
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accesso negato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========================================
  # ANALYTICS
  # ========================================
...